+++ src/bundle-orchestrator.ts
@@ -1,6 +1,6 @@
 import Vinyl from "vinyl";
 import { Transform, TransformCallback } from "stream";
-import { Logger, dummyLogger } from "ts-log";
+import TsLog from "ts-log";
 import { Config } from "./config/config.js";
 import extend from "just-extend";
 import { resolve as resolvePath } from "path";
@@ -59,7 +59,7 @@ function bundleFactory(
     rawPaths: string[],
     cwd: string,
     joiner: BundleStreamFactory,
-    logger: Logger
+    logger: TsLog.Logger
 ) {
     const paths = [];
     for (const rawPath of rawPaths) {
@@ -114,7 +114,7 @@ export class BundleOrchestrator extends Transform {
 
     private resultsCallback?: ResultsCallback;
 
-    private logger: Logger = dummyLogger;
+    private logger: TsLog.Logger = TsLog.dummyLogger;
 
     /**
      * @param config - Raw (but valid) configuration file used for bundle resolution.
@@ -127,7 +127,7 @@ export class BundleOrchestrator extends Transform {
         });
 
         // First up, we assign the logger if its there
-        /* istanbul ignore else */
+        /* c8 ignore else */
         if (config.Logger) this.logger = config.Logger;
 
         // Results callback
@@ -189,10 +189,9 @@ export class BundleOrchestrator extends Transform {
 
             callback();
         }
+        /* c8 ignore next 4 */
         catch (error) {
-            /* istanbul ignore next This is only here to ensure we are a polite pipeline member, and cannot be reasonably tested */
             this.logger.error("_transform completed with error", { error });
-            /* istanbul ignore next */
             callback(new PluginError(PluginName, error));
         }
     }