+++ src/config/merge-configs.ts
@@ -1,7 +1,7 @@
 import { Config } from "./config.js";
-import MergeBundle from "./merge-bundle.js";
+import { MergeBundle } from "./merge-bundle.js";
 import extend from "just-extend";
-import ExtendedError from "errlop";
+import ono from "@jsdevtools/ono";
 
 /**
  * Merges a collection of configurations.
@@ -11,7 +11,7 @@ import ExtendedError from "errlop";
  * @param rawConfigs - Raw (untransformed) configurations to merge.
  * @public
  */
-export default function MergeConfigs(rawConfigs: Config[]): Config {
+export function MergeConfigs(rawConfigs: Config[]): Config {
     // No point doing processing if we've got only 1 item
     if (rawConfigs.length === 1) return rawConfigs[0];
 
@@ -35,7 +35,7 @@ export default function MergeConfigs(rawConfigs: Config[]): Config {
                         nextConfig.bundle[bundleName] = MergeBundle(outConfig.bundle[bundleName], nextConfig.bundle[bundleName]);
                     }
                     catch (exception) {
-                        throw new ExtendedError(`Exception raised while merging bundle '${bundleName}' in the raw configuration at index '${rawConfigs.indexOf(config)}'.`, exception);
+                        throw ono(exception, `Exception raised while merging bundle '${bundleName}' in the raw configuration at index '${rawConfigs.indexOf(config)}'.`);
                     }
                 }
                 // Otherwise just set it