+++ src/main.ts
@@ -1,17 +1,12 @@
-export { default as MergeRawConfigs } from "./config/merge-configs.js";
-export { default as ValidateRawConfig } from "./config/validate-config.js";
+export { MergeConfigs as MergeRawConfigs } from "./config/merge-configs.js";
+export { ValidateConfig as ValidateRawConfig } from "./config/validate-config.js";
 export {
-    /**
-     * @deprecated
-     */
-    BundleOrchestrator as BundleOrchastrator,// spell-checker:disable-line
+    BundleOrchestrator as default,
     BundleOrchestrator,
     Bundlers,
     Results,
     ResultsCallback,
 } from "./bundle-orchestrator.js";
-import { BundleOrchestrator } from "./bundle-orchestrator.js";
-export default BundleOrchestrator;
 export { BundleStreamFactory } from "./bundle";
 export {
     Bundle,